refactor: 스터디 디테일 컴포넌트 폴더 내 스터디 메인 서브 폴더 생성
