+++ src/components/study-detail/main/ManageNoticeModal.tsx
@@ -1,9 +1,9 @@
-import ModalOverlay from '../ModalOverlay';
-import Button from '../Button';
-import CloseIcon from '../../assets/icons/close.svg';
-import { handleMaxLengthChange } from '../../utils/commonUtils';
-import NoticeIcon from '../../assets/icons/notice.svg';
-import useManageModal from '../../hooks/study-detail/useManageModal';
+import ModalOverlay from '../../ModalOverlay';
+import Button from '../../Button';
+import CloseIcon from '../../../assets/icons/close.svg';
+import { handleMaxLengthChange } from '../../../utils/commonUtils';
+import NoticeIcon from '../../../assets/icons/notice.svg';
+import useManageModal from '../../../hooks/study-detail/useManageModal';
 
 export default function ManageNoticeModal({
   isOpen,
