+++ src/components/study-detail/main/Notice.tsx
@@ -1,7 +1,7 @@
 import { overlay } from 'overlay-kit';
 import NoticeModal from './NoticeModal';
-import useGetNotice from '../../hooks/study-detail/useGetNotice';
-import NoticeIcon from '../../assets/icons/notice.svg';
+import useGetNotice from '../../../hooks/study-detail/useGetNotice';
+import NoticeIcon from '../../../assets/icons/notice.svg';
 
 export default function Notice({ studyId, hasNotice }: { studyId: number; hasNotice: boolean }) {
   const { notice, isLoading } = useGetNotice(studyId, hasNotice);
