+++ src/components/study-detail/main/TodoListSection.tsx
@@ -1,6 +1,6 @@
-import useTodoList from '../../hooks/study-detail/todo/useTodoList';
-import PlayIcon from '../../assets/icons/play.svg';
-import CloseIcon from '../../assets/icons/close.svg';
+import useTodoList from '../../../hooks/study-detail/todo/useTodoList';
+import PlayIcon from '../../../assets/icons/play.svg';
+import CloseIcon from '../../../assets/icons/close.svg';
 
 export default function TodoListSection({ studyId }: { studyId: number }) {
   const {
